requirements.txt
flufl.enum
rdflib
wrapt
inflection